+++ src/addCard/components/MonthInput.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { validator, checkMaxLength, checkIsNaN, checkRange } from '../validator';
-import { MAX_LENGTH, MIN_LENGTH, RANGE } from '../constants';
+import { validator, checkMaxLength, checkIsNaN, checkRange } from '../../validator';
+import { MAX_LENGTH, MIN_LENGTH, RANGE } from '../../constants';
 import Input from './Input';
 
 function MonthInput({ value, name }) {
