+++ src/addCard/storybook/AddCard.stories.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import AddCard from '../components/AddCard';
-import '../css/index.css';
-import '../css/App.css';
+import '../../css/index.css';
+import '../../css/App.css';
 
 export default {
   title: 'AddCard',
